Add METplus demo code.
This is the METplus demo that Rachel has been working on. I'm not sure that
this will be its final location in the codebase, and it needs a lot of
generalising to turn it into a task (most paths are still hardcoded, for
example), but it is a good place to start.