Add tomli as explicit dependency
It is needed to support reading TOML in python < 3.11. It was being pulled in
by the build system before, but as it is a runtime dependency it should be
xplicitly listed.

Also made the import exception more specific.

Add some explicit exceptions to flake8 checks

Add newline to satisfy pre-commit