Rewrite operator recipe parser
Partly refactored to avoid exec. This is only halfway complete, and doesn't
actually run. I'll pick it up on Monday and this commit will be squashed with
the next one.

Add helper function to get an operator from its name

Allow operators to take arbitrary keyword arguments

Rewrite operator recipe parser

This rewrite avoids using exec, so should be generally immune to arbitrary code
execution. It also is a much cleaner and simpler implimentation.

Replace match-case and add docstring in get_operator

match-case syntax was added in 3.10, and as we are aiming to work with 3.8 we
can't use it.