Reorganise documentation to better separate different kinds
The documentation has been reorganised to better allow distinct
kinds of documentation, with distinct audiences, to be written.

This involves splitting the documentation into several distinct
sections based on how it should be written. This is Tutorials,
How-to guides, References, and Explanations.

 * Tutorials are for people learning their way around.
 * How-to guides answer the "how?".
 * References answer the "what?".
 * Explanations answer the "why?".

This is a commonly used, and well regarded documentations style used
by projects such as NumPy.

More information can be found here: https://documentation.divio.com/