Add recipe variables
These are variables that can be put into a recipe and are replaced
at runtime with values specified as command line arguments.

In a recipe file they are any string value of the form `$VARIABLE`,
where the variable name can consist of A-Z and _.

The runtime value is given to cset bake in the form `--VARIABLE=value`.

Add fix variable handling and add tests

Test for missing variable value

Add documentation for recipe variables