Cleanup run_cset_recipe app
This is the result of an effort to make the plot ID deterministically
generated though it became apparent this wasn't the right place to implement
it.

Make plot ID deterministically generated

The plot ID is now the SHA256 hash of the recipe. As this is deterministic
it will allow updating plots in other cycles.

Revert to UUID for plot ID as hashing isn't working

We can't hash the recipe there as the templating hasn't yet happened.