Copy files within source directory when installing locally
This removes a presumption that the source directory is always called "CSET".

We expand out variables and copy files within python. This avoids shelling
out a subprocess.

Also, don't try to install cset dependencies with pip. They are already
contained within the conda environment, so it is always wasted effort.

Also fix the stdout decoding to use the correct encoding. It was
previously coincidentally correct.

Fixes #471