Ignore cartopy DownloadWarning in pytest
As CI starts from a fresh slate each time we won't yet have the resources
downloaded. We do cache them, but caches are only per-branch, and thus we
still need to download on the first run per branch.

Also removed the dateutil warning filter, as that issue has been fixed.