Merge install_local_cset into build_conda
The install_local_cset script was converted to bash as part of this.

This means we only have a single job for environment setup, and test
that the environment actually works as part of that.

We also skip copying non-git hidden files when installing local CSET,
as they can often include large things like conda environments that
slow the script significantly.