Preload plot images on web page
This ensures that the slider slides smoothly, as the image can be served
from the browser cache.

I used JavaScript rather than <link rel="preload" as="image" href="...">
as it was much easier to implement, as we don't currently have any way to
loop over the plots at template time. If this changes we may want to
switch, as HTML is both more explicit, and also taxes the browser less
due to not having to create a non-displayed image element. The browser
also seems to have more conservative limits for rel=preload, so it won't
do as many concurrent requests.

Fixes #780