Make area constraint much stricter, and handle date line
The operator now requires all arguments to either be numbers, or all be None
if you don't want to constrain the area.

The test has been mildly improved, so at least it checks for both of the
named functions on the coordinates, though it still doesn't test that they
do anything.