CLI/read operator data loading improvements
This handles globbing anywhere in the input path, moves to treating the
provided patterns as strings rather than paths before they have been
resolved, and generally makes the path and file loading more
consistently handled.

Also included is marking the first path loaded as comparison base, which
will be used by difference operators to know what to compare against.
In practice this will translate to the first model in the GUI being used
as the base.