Generate rose-meta.conf
This change switches to a generated rose-meta.conf, so it is much easier
to maintain. It also removes a number of obsolete keys, updates descriptions,
makes analysis_offset and analysis_period per-model, and removes
the s/crimes against rose edit/mapping between model and fieldname/.

One of the biggest changes is the move to a simple list for surface fields
instead of the complicated 2D affair we had before.