Generate rose-meta.conf
This change switches to a generated rose-meta.conf, so it is much easier
to maintain. It also removes a number of obsolete keys, updates descriptions,
makes analysis_offset and analysis_period per-model, and removes
the s/crimes against rose edit/mapping between model and fieldname/.

One of the biggest changes is the move to a simple list for model fields
instead of the complicated table affair we had before.

Fixes #667
Fixes #808
Fixes #817
Fixes #859

Update example rose-suite.conf