Fix fencepost problem with colorbar bins
This makes the numbers on the colorbar labels much rounder if the
range is a multiple of 20.

This comes from the fencepost problem, which describes how you need
n+1 dividers to support n bins. For example, imagine a fence with 5
segments, like so:

    |###|###|###|###|###|

Notice that we have 6 posts (denoted with | characters) for our 5
fence segments. This is the fence post problem.