Always build conda environment within workflow share dir
By always building the conda environment within the workflow share
directory we are able to remove several bits of configuration, and
simplify the environment creation scripts.

We will be paying the cost of somewhat slower builds, but we should gain
increased reliability, and more robustness when running multiple
simultaneous runs.