Avoid writing meta.json outside tempdir during test
The call to get_metadata would prompt for the creation of a meta.json
file, but this test was running outside of a temporary directory, so the
meta.json was left in your current working directory after the test had
run. The test now runs in a temporary directory, avoiding this issue.