Plot multiple models on histogram plots
With this change multiple models will be plotted as separate lines on a
shared histogram axes. If the model's name is specified it is used to
consistently assign a line colour and shown in a legend on the plot.

As we determined that almost all users would want this, there is no
option to plot the models separately. As a workaround CSET can be run
multiple times, but if there is significant interest we could add a
preference for this.

Fixes #566, #1110, #1307