Add cset parbake command
This command takes a recipe and its variables, and produces a recipe
with all those variables templated in. This allows subsequent baking
without any additional context beyond the recipe file.

The existing cset bake command can continue to be used, but now may also
be used to run the parbaked recipes. A new workflow script has been
included to do this for many recipes in a single cylc job, which should
lead to significant speedups.