Remove build-conda task
By bringing in the environment from outside the task we no longer need
to build the conda enviornment at the start of each run. Additionally we
can just as easily use a development version as a released version.
Therefore we are able to remove the USE_LOCAL_CSET settings.

A basic environment validation step is added at the start of the
workflow, to aid debugging.