Set 20 minute runtime limit for GitHub Actions jobs
This will prevent us from wasting too many minutes when a job gets
stuck, and hopefully alert us to the situation faster.

I've chosen 20 minutes as it is short enough to avoiding wasting too
many minutes, compared to the default 6 hours, while it is long enough
to avoiding having to change the limit for most jobs.

If a job does need longer than 20 minutes, the limit can be increased on
a per-job basis.