Skip posting coverage report on PRs from forks
For security reasons pull requests from forks cannot modify the upstream repository content, including the comments on its own PR. This causes our code coverage check to fail on PRs from forks as it fails to post the comment.

To avoid this we skip posting the comment if it fails, allowing the check to complete successfully after checking the minimum coverage has been met.

Fixes #578