Move workflow utils into rose apps
As the workflow is now inside the package, its scripts are able to be
tested in place.

The simple, self-contained scripts were just moved over, while the more
complex fetch_data script needed to import the fetch_data infrastructure
available to the other scripts, but given they are in the same directory
we can rely in the current directory being added to the PYTHONPATH.

While the scripts where being moved I took the opportunity to add licence
headers where they were missing.

The removes the last file in _workflow_utils, so that subpackage is
deleted.