Remove preprocess task
It was only ever a performance hack, and it is causing issues when
loading ensemble data. Therefore it is removed.

There was one piece of functionality that needed to be retained, which
is the removal of the output at T=0. While I think it should probably
also be removed, that will be a separate issue as it requires fixing the
underlying cause for its existence.

Fixes #1511