Simplify release instructions to reflect bundled workflow
As the workflow is now included in the CSET package, we no longer need
to do the awkward dance of publishing the package to conda-forge, then
updating the lockfiles after a painful wait, and finally uploading a
manually created tarball.

I'd still like to further automate this process, especially around the
creation of the changelog element of the release notes, and updating the
conda-forge feedstock, but it is already a significant step up from what
we had before.