Add bake_recipes rose app
This uses rose-bunch to bake multiple recipes in parallel inside of a
single cylc job. This makes it significantly faster than running
thousands of tiny jobs.

To allow it to dynamically find the available jobs, cylc executes the
baker.sh script, which writes a rose optional configuration specifying
the recipes and parallelism, then runs the rose app using that
configuration.