WIP Add search capability to CSET UI
Additionally included is a conversion to JSON Lines (In theory it should
be more efficient, but in practice we still load the whole string first,
so I may revert. At least it looks nicer in the index file.)

Finally there are also some experiments with the UI for facets, ahead of
getting feedback from the user interviews.

The framework is there for facet-based searching, but the current
implementation of search is very basic, only doing exact substring
matching. Additional functionality should be added to the test function
constructed in parse_query.