Avoid extracting recipes files if we don't need to
Previously we always got a file handle to the recipe file ahead of
checking whether it already existed, which caused the files to be
extracted from the python package into a temporary location, even if we
were then not going to use them.

We now only get the file handle after checking that the recipe doesn't
already exist on disk, which should make this significantly faster.

We also now only log skipped recipe unpacking at DEBUG log level, as it
was a bit spammy before (and doubles our required IO!)