Use rose-bunch incremental mode when baking
This is the default, but it has been made explicit to be sure it is
enabled.

There are two issues fixed in this commit. The first is that we were
using the CYLC_TASK_TRY_NUMBER variable to determine if we were
retrying, which gets reset on a manual cylc trigger.
CYLC_TASK_SUBMIT_NUMBER is not, so that all runs after the first will
now run serially, regardless of how they were triggered.

The other issue was that rose_bunch was never running tasks
incrementally due to us changing the parallelism in the configuration.
We now avoid this by updating the recorded configuration in rose_bunch's
database to match.

This is decidedly a hack, but seems to work well, and means that we now
only retry the failed tasks, as intended. It also adds a dependency on
the sqlite command line, which will almost certainly be available, but
should be documented to be sure.