Use long_name for difference cube over standard_name
As we match colourbars off of names, we need to use them consistently.
We have decided in other parts of the codebase to treat the long_name as
higher priority than the standard_name, but this was missed in the
difference operator, resulting in some difference plots not picking up
their colourbars.

Fixes #1710